# Quantization improvements