@compilade 2025-03-28
Nice! It seems like your improved `make_qx_quants` is extremely similar to `make_qkxh_quants` when starting the search from `MIN(abs(nmin), abs(nmax)) - 1` instead of `MIN(abs(nmin), abs(nmax)) / 2` (when comparing the equirectangular projections). This would also make `make_qkxh_quants` faster (though I don't know by how much).

Here's your improved `make_qx_quants` with settings from `Q4_0`:

![equirectangular-tmp-2048](https://github.com/user-attachments/assets/3b0c3d0e-92c7-43f9-b498-2bb3adf4143c)

And your improved `quantize_row_iq4_nl_impl` looks like this:

![equirectangular-tmp2-2048](https://github.com/user-attachments/assets/855d814b-15bd-46b8-8546-42ed2f71f4b5)


Very interesting approach with the gradient.