@ikawrakow 2025-03-28
And what are the two coordinates of the plot? I understand it is a projection, but what is it that is being projected?

> Very interesting approach with the gradient.

That would be the standard way to approach an optimization problem, no?