@compilade 2025-03-28
> And what are the two coordinates of the plot? I understand it is a projection, but what is it that is being projected?

The horizontal coordinates is `theta` which goes from 0 to 2*π radians, while the vertical coordinates is `phi`, which goes from 0 to π radians.

The vectors tested have the form $[\sin(\phi) \cdot \cos(\theta), \sin(\phi) \cdot \sin(\theta), \cos(\phi)]$

The script which I'm using is <https://github.com/compilade/rounding-experiments/blob/main/equirectangular.py>, although I have some local modifications to make it use other rounding algorithms, which are defined in [`rounding-impl.c`](https://github.com/compilade/rounding-experiments/blob/main/rounding-impl.c) with Python bindings in [`rounding_c.py`](https://github.com/compilade/rounding-experiments/blob/main/rounding_c.py).

> > Very interesting approach with the gradient.
> 
> That would be the standard way to approach an optimization problem, no?

Sure. Being standard doesn't mean it's not interesting. You have made the gradients explicit, which I appreciate.

And ***if*** your gradient search and my cumulative search (once the range is reduced) are equivalent (or close enough), that in itself is interesting, since I did not explicitly use gradients.

I really like when different approaches end up being equivalent (or close enough) because this makes them easier to understand, explain and generalize to other cases (notably, my approach might be harder to adapt to grid-restricted i-quants).

(If they are not equivalent, this is still very cool, even if this is technically using a standard approach)

I will compare the speed and perplexity of narrower cumulative search with this once I have some spare time, since I do think reducing the searched range will greatly improve the speed of my (currently quite slow) proposed algorithms.